+++ apps/OboeTester/app/src/main/cpp/NativeAudioContext.h
@@ -46,7 +46,7 @@
 
 #include "FullDuplexAnalyzer.h"
 #include "FullDuplexEcho.h"
-#include "FullDuplexStream.h"
+#include "oboe/FullDuplexStream.h"
 #include "analyzer/GlitchAnalyzer.h"
 #include "analyzer/DataPathAnalyzer.h"
 #include "InputStreamCallbackAnalyzer.h"
