+++ include/oboe/Oboe.h
@@ -35,5 +35,6 @@
 #include "oboe/StabilizedCallback.h"
 #include "oboe/FifoBuffer.h"
 #include "oboe/OboeExtensions.h"
+#include "oboe/FullDuplexStream.h"
 
 #endif //OBOE_OBOE_H
